Upgrade debugging tools
Swap out pry and byebug for pry-byebug which gives all of the
capabilities of byebug (step-by-step and stack navigation) inside of the
pry CLI.

Add pry-rescue which can trigger pry automatically on any unhandled
exception (or test failure when combined with minitest).

Remove explicit dependency on binding_of_caller because it's already
included by other dependencies which we use more explicitly.