Refactor UserInvite's role validation test
Previously we were assuming that a single fixture could be used with any
valid role and be considered valid. Now we're defining separate fixtures
for each role which are valid for that role.

The net effect of the test is still the same, we're verifying that there
exists at least one valid state for every known role, and that an
otherwise valid state with the role not set to a known role is invalid.