Handle unknown repo case for group projects
When GitHub has a PR made from a repository/fork that has been deleted
it shows "unknown repository" as the source on the [PR page][1] and the
API does not include any repo information for the "head" section of the
PR data.

We now check whether that information is available from the API and
effectively ignore the PR if it does not exist. My understanding is that
this means the submission line for each associated student will continue
to be listed as unsubmitted, however I do not have a good test case for
this situation.

This fix only applies to the logic for group projects as the individual
project does not require the same repo information.

Note: The safety check code that I've put in the `contributors_url`
method could be cleaned up a bit if we switch to Ruby 2.3.0 or later,
which supports `Hash#dig` for exactly this purpose.

[1]: https://github.com/Ada-C7/VideoStoreConsumer/pull/8