Only redirect to PR index if user can view PRs
The check for `current_user` is no longer correct with the new
authorization system because there are possibly logged in user accounts
who do not have access to view PRs.

When that happens the PR index action redirects them back to the `home`
route, which redirects to the PR index, causing an infinite loop.