Editing/destroying a Repo no longer raises error
This was caused by an attempt to fix the error in `ReposController#show`
which was too simple. CanCanCan's `load_and_authorize_resource` method
attempts to determine which param value to use for loading a given
resource based on a few factors.

In this case we need to be explicit that we want to load the `Repo`
object as a parent (i.e. with the `repo_id` param) and load the `Cohort`
object not as a parent (with the `id` param). The routes definition has
been updated to match these param names as well.

However, for all other actions in `ReposController` we need to load the
`Repo` as though it is not a parent (using `id` param), so we have a
separate specification for that case.

Because the `Cohort` object is being loaded by CanCanCan we no longer
need the loading logic in the `show` action, and instead all resource
loading failures trigger the `rescue_from` block (whose message text has
been genericized).

Finally, we now test what happens when the `show` action is called with
an invalid cohort ID.

NOTE: This could all probably be simplified in the future by adding a
`Project` model which works as a join table between `Repo` and `Cohort`.
There are other reasons for implementing this, but it has a side benefit
that we won't need this one-off controller action which uses nested
resources when all other actions do not.