Cleanup DB seeding logic and data
We now seed data that matches actual GitHub data (instructors, students,
and repositories). We also no longer attempt to seed students from a CSV
file, as that should be done manually and only for live production
deployments.

Note: We seed a single user, tied to the GitHub account
"ada-instructor-1". This is a real GitHub account and if you have the
password then you can use it to log in and invite your personal GitHub
account as an instructor.

If you do not have the password you will need to log in with your
personal GitHub account, then open `rails console` and directly set your
`User` record to have the `instructor` role.