OmniAuth mocks have correct provider in all cases
This was a long-standing bug in our code for creating OmniAuth mocks,
because I did not understand that OmniAuth only supports creating a
single mock per provider.

We use multiple mocks for the GitHub provider to represent different
users logging into the app. In those cases the "name" of the mock was
being used as the provider, for example "uninvited" instead of "github".

We never encountered this issue in previous testing because we did not
have any tests which logged in using one mock and then logged in a
second time using a different mock.

We now force every mock to use the same provider as the default mock,
which will probably always be "github" for this project.