User info is updated from GitHub data on login
This is generally good practice. For example if someone updates their
name on the GitHub website they should see that reflected in our app.

But it's also necessary for us because we've added the GitHub login
field to our User account and that needs to be updated for any existing
accounts which do not have it stored yet.