Upgrade omniauth gem to 1.8.1
Originally I upgraded to 1.3.2 because GitHub informed us of a security
vulnerability with our existing version (1.3.1).

However, that upgrade also resulted in upgrading the hashie gem's
version which further resulted in a warning message upon deployment
because of changes in the hashie gem.

Upgrading to a more recent version of omniauth is apparently the fix, so
this is upgrading to the most recent stable version.